scripts/requirements.txt
yq
python-frontmatter
Markdown
GitPython
requests
ruamel.yaml
charset-normalizer
